🎨 Make `sphnx_link_by_labels.ext` Sphinx ext
This patch provides a demonstration of a Sphinx extension that
collects document labels provides through a `labels` field in
file-wide metadata that Sphinx supports out of the box[[1]].

The extension, then, provides a directive for displaying all the
document names with labels requested through its arguments.

It uses some of the techniques documented in the official TODO list
tutorial in the Sphinx documentation[[2]].

This extension relies to the built-in hook points, implemented as
events in Sphinx[[3]]. It first transforms the labels that Sphinx
collected into an efficient mapping when reacting to the `env-updated`
event and saves it in the environment object.
The directive essentially only injects the "request" nodes with label
lists into the doctree (this happens around `doctree-read`, way
before `env-updated`, so before app the documents get parsed).
Later in the lifecycle, the extension reacts to the `doctree-resolved`
that is invoked for each document, retrieves the document names from
the mapping created earlier, and then, replaces all the temporary
"request" nodes (since the directive could be used a few times per
document) with text listing the document names.

[1]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/field-lists.html
[2]: https://www.sphinx-doc.org/en/master/development/tutorials/todo.html
[3]: https://www.sphinx-doc.org/en/master/extdev/appapi.html#sphinx-core-events